Cleaned up Redux state / auth actions
Moved transient state (i.e. error message and login loading status) out
of Redux and into login component state. Also consolidated
actions/login.js and actions/logout.js into actions/auth.js.
