+++ src/client/components/sidebar.jsx
@@ -14,7 +14,7 @@ import {
   ADMIN_PANEL_PAGE_ROUTE,
   NOT_FOUND_DEMO_PAGE_ROUTE,
 } from '../../shared/routes';
-import { logoutUser } from '../actions/logout';
+import { logoutUser } from '../actions/auth';
 import styles from '../styles/components/sidebar.m.scss';
 
 type Props = {
